editor/libeditor/tests/test_bug366682.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=366682
-->
<head>
  <title>Test for Bug 366682</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script src="spellcheck.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=366682">Mozilla Bug 366682</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 366682 **/

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

var gMisspeltWords;

function getEdit() {
  return document.getElementById('edit');
}

function editDoc() {
  return getEdit().contentDocument;
}

function getEditor() {
  var Ci = Components.interfaces;
  var win = editDoc().defaultView;
  var editingSession = win.QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIWebNavigation)
                          .QueryInterface(Ci.nsIInterfaceRequestor)
                          .getInterface(Ci.nsIEditingSession);
  return editingSession.getEditorForWindow(win);
}

function runTest() {
  editDoc().body.innerHTML = "<div>errror and an other errror</div>";
  gMisspeltWords = ["errror", "errror"];
  editDoc().designMode = "on";

  Components.utils.import("resource://gre/modules/AsyncSpellCheckTestHelper.jsm");
  onSpellCheck(editDoc().documentElement, evalTest);
}

function evalTest() {
  ok(isSpellingCheckOk(getEditor(), gMisspeltWords),
     "All misspellings accounted for.");
  SimpleTest.finish();
}
</script>
</pre>

<iframe id="edit" width="200" height="100" src="about:blank"></iframe>

</body>
</html>
